rename package of FabricForDI
