+++ src/main/java/executor/service/ObjectFactory/ObjectFactory.java
@@ -1,4 +1,4 @@
-package executor.service.DIFactory;
+package executor.service.ObjectFactory;
 
 public interface ObjectFactory {
     <T> T create(Class<T> clazz) ;
